requirements.txt
robotframework
robotframework-requests
robotframework-jsonschemalibrary
pyyaml